поиск не работает

+++ RGR/tasks/task_2/shortest_way.h
@@ -54,7 +54,7 @@ public:
         search();
     }
 
-    [[nodiscard]] int ST(const int v) const {
+    [[nodiscard]] unsigned long ST(const unsigned long v) const {
         return st[v];
     }
 };
